* commit '1ffc25c7d1b850d9c934d4357094ef902c7c913a':
  NM: comments and minor cleanup

+++ src/common/point_xyz.cpp
@@ -55,13 +55,13 @@ PointXYZ operator-(PointXYZ const & p)
 
 float norm(PointXYZ const & p)
 {
-  return sqrtf(p.x*p.x + p.y*p.y + p.z*p.z);
+  return std::sqrt(p.x*p.x + p.y*p.y + p.z*p.z);
 }
 
 
 PointXYZ normalize(PointXYZ const & p)
 {
-  float norm = sqrtf(p.x*p.x + p.y*p.y + p.z*p.z);
+  float norm = std::sqrt(p.x*p.x + p.y*p.y + p.z*p.z);
 
   float s = 1.0f / norm;
